refactor(cas): review fixes — plain postgres image, health module
Use postgres:18 instead of the alpine image, drop the data-path comment,
and move the health endpoint out of main.rs into its own module.
